remove swift sources
The extensions in OCResult2Swift.swift caused build errors because it tries to access generic state type from an untyped extension (swift extensions cannot be generic either). These would need to be rewritten as global functions that take an object of one type and returns the other, instead of being instance methods on the now generic objc class.